src/notification_service/requirements.txt
flask
kafka-python
#smtplib
